backend/requirements.txt
Django==2.2.4
django-crispy-forms==1.7.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.7.3.1
structlog==19.1.0
supervisor==4.0.2
requests
python-dateutil==2.8.0